pyproject.toml
[tool.poetry]
name = "todoapi"
version = "0.1.0"
description = ""
authors = ["UniqueMaharjan <unique.maharzn@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.2.0"
django-cors-headers = "^3.13.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
